Full rewrite (#64)
## Full rewrite of this crate

The main problem was that the API followed the implementation and not the other way round. Some API improvements weren't possible because they weren't implementable.

With the new implementation, several API decisions that bothered me were finally improved.

### Breaking Changes
- Remove `Toplevel::nested` (which was a rather ugly hack), replace with configurable `ErrorAction`s.
- Remove partial shutdown mechanism; replace with `NestedSubsystem::initiate_shutdown()` and `NestedSubsystem::join()`.
- Add `SubsystemHandle::wait_for_children()`.
- Remove `Toplevel::start`. Instead, add a single subsystem to `Toplevel::new`, which can then spawn all further subsystems.
- Rework `SubsystemHandle::start`; it now takes a `SubsystemBuilder` object that provides further subsystem configuration options.
- Minimum Supported Rust Version (MSRV) is now `1.63.0`.


The rest of the API is almost identical and migration should be mostly straight forward.